+++ base/gsfunc.c
@@ -1,4 +1,4 @@
-/* Copyright (C) 2001-2019 Artifex Software, Inc.
+/* Copyright (C) 2001-2020 Artifex Software, Inc.
    All Rights Reserved.
 
    This software is provided AS-IS with no warranty, either express or
@@ -127,7 +127,7 @@ fn_copy_values(const void *pvalues, int count, int size, gs_memory_t *mem)
         void *values = gs_alloc_byte_array(mem, count, size, "fn_copy_values");
 
         if (values)
-            memcpy(values, pvalues, count * size);
+            memcpy(values, pvalues, (size_t)count * size);
         return values;
     } else
         return 0;		/* caller must check */
