+++ base/gxpcache.h
@@ -1,4 +1,4 @@
-/* Copyright (C) 2001-2019 Artifex Software, Inc.
+/* Copyright (C) 2001-2020 Artifex Software, Inc.
    All Rights Reserved.
 
    This software is provided AS-IS with no warranty, either express or
@@ -35,8 +35,8 @@ struct gx_pattern_cache_s {
     uint num_tiles;
     uint tiles_used;
     uint next;			/* round-robin index */
-    ulong bits_used;
-    ulong max_bits;
+    size_t bits_used;
+    size_t max_bits;
     void (*free_all) (gx_pattern_cache *);
 };
