+++ devices/gdevdm24.c
@@ -1,4 +1,4 @@
-/* Copyright (C) 2001-2019 Artifex Software, Inc.
+/* Copyright (C) 2001-2020 Artifex Software, Inc.
    All Rights Reserved.
 
    This software is provided AS-IS with no warranty, either express or
@@ -150,7 +150,7 @@ dot24_print_page (gx_device_printer *pdev, gp_file *prn_stream, char *init_strin
             if (code == 0)
               {
                 /* Pad with lines of zeros. */
-                memset (inp, 0, (24 - lcnt) * line_size);
+                memset (inp, 0, (size_t)(24 - lcnt) * line_size);
                 break;
               }
           }
@@ -162,7 +162,7 @@ dot24_print_page (gx_device_printer *pdev, gp_file *prn_stream, char *init_strin
             if (code == 0)
               {
                 /* Pad with lines of zeros. */
-                memset (inp, 0, (24 - lcnt) * line_size);
+                memset (inp, 0, (size_t)(24 - lcnt) * line_size);
                 break;
               }
           }
